fix: align with TagV2 token structure fix the token structure and types
